fix all the peotry issues by moving code to bee-hive cli directory. Now run CLI with: poetry run ./beeAI -h You can run the validate command with: poetry run ./beeAI validate ../tools/workflow_schema.json ../test/schema/simple_workflow.yml
