+++ bee-hive/beeAI
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python3 ./cli/beeAI.py $@
