+++ bee-hive/pyproject.toml
@@ -11,6 +11,8 @@ python = ">= 3.11, < 3.13"
 pyyaml = "^6.0.2"
 openai = "^1.56.2"
 python-dotenv = "^1.0.1"
+docopt = "^0.6.2"
+jsonschema = "^4.23.0"
 
 
 [tool.poetry.group.dev.dependencies]
