--- beeAI
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-python3 ./tools/cli/beeAI.py $@
