+++ tools/workflow_schema.json
@@ -26,7 +26,7 @@
         }
       },
       "required": [
-	"name"
+    "name"
       ]
     },
     "spec": {
@@ -54,37 +54,37 @@
                 "labels": {
                   "type": "object",
                   "description": "workflow labels, key: value pairs"
-		}
+        }
               }
             }
-	  }
+      }
         },
         "agents": {
           "type": "array",
           "items": {
-	    "type": "string"
+        "type": "string"
           }
         },
         "exception": {
-	  "type": "string"
+      "type": "string"
         },
         "prompt": {
-	  "type": "string"
+      "type": "string"
         },
-	"steps": {
+    "steps": {
           "type": "array",
           "items": {
             "type": "object",
             "properties": {
-	      "name": {
+          "name": {
                 "type": "string",
                 "description": "step name"
               },
-	      "agent": {
+          "agent": {
                 "type": "string",
-		"description": "the agent for this step"
+        "description": "the agent for this step"
               },
-	      "condition": {
+          "condition": {
                 "type": "array",
                 "description": "if/then/else or case/do/default condition",
                 "items": {
@@ -101,10 +101,10 @@
                     },
                     "case": {
                       "type": "string"
-	            },
+                },
                     "do": {
                       "type": "string"
-	            },
+                },
                     "default": {
                       "type": "string"
                     }
